+++ langworld_db_data/removers/listed_value_remover.py
@@ -7,7 +7,7 @@ from langworld_db_data.constants.literals import (
     KEY_FOR_VALUE_TYPE,
 )
 from langworld_db_data.tools.files.csv_xls import read_dicts_from_csv, write_csv
-from langworld_db_data.tools.value_ids.value_ids import extract_feature_id, extract_value_index
+from langworld_db_data.tools.value_ids.extract import extract_feature_id, extract_value_index
 
 
 class ListedValueRemoverError(Exception):
