+++ langworld_db_data/tools/files/csv_xls.py
@@ -15,7 +15,7 @@ import openpyxl.worksheet.worksheet
 from openpyxl import load_workbook
 
 from langworld_db_data.constants.literals import ID_SEPARATOR
-from langworld_db_data.tools.value_ids.value_ids import (
+from langworld_db_data.tools.value_ids.extract import (
     extract_feature_index,
     extract_value_index,
 )
