+++ tests/test_tools_value_ids_value_ids.py
@@ -1,4 +1,4 @@
-from langworld_db_data.tools.value_ids.value_ids import (
+from langworld_db_data.tools.value_ids.extract import (
     _split_value_id,
     extract_category_id,
     extract_feature_id,
